requirements.txt
numpy 
pandas
scikit-learn
joblib
seaborn
ipykernel
fastapi
flask
mlflow
notebook
matplotlib
python-box
pyYAML
tqdm
ensure
types-pyYAML
Flask
Flask-Cors
mlflow
dvc
python-dotenv
dagshub
dvc-s3
uvicorn
python-multipart
pydantic
streamlit
ensure
box